Move ValueClassUnboxBoxConverter to use serialization and deserialization
