+++ src/main/kotlin/io/github/projectmapk/jackson/module/kogera/Converters.kt
@@ -1,6 +1,8 @@
 package io.github.projectmapk.jackson.module.kogera
 
+import com.fasterxml.jackson.databind.JavaType
 import com.fasterxml.jackson.databind.ser.std.StdDelegatingSerializer
+import com.fasterxml.jackson.databind.type.TypeFactory
 import com.fasterxml.jackson.databind.util.StdConverter
 
 /**
@@ -21,4 +23,14 @@ internal class ValueClassBoxConverter<S : Any?, D : Any>(
     override fun convert(value: S): D = boxMethod.invoke(null, value) as D
 
     val delegatingSerializer: StdDelegatingSerializer by lazy { StdDelegatingSerializer(this) }
+}
+
+internal class ValueClassUnboxConverter<T : Any>(private val valueClass: Class<T>) : StdConverter<T, Any?>() {
+    private val unboxMethod = valueClass.getDeclaredMethod("unbox-impl").apply {
+        if (!this.isAccessible) this.isAccessible = true
+    }
+
+    override fun convert(value: T): Any? = unboxMethod.invoke(value)
+
+    override fun getInputType(typeFactory: TypeFactory): JavaType = typeFactory.constructType(valueClass)
 }
