+++ src/main/kotlin/io/github/projectmapk/jackson/module/kogera/deser/Converters.kt
@@ -10,16 +10,6 @@ import io.github.projectmapk.jackson.module.kogera.JavaDuration
 import io.github.projectmapk.jackson.module.kogera.KotlinDuration
 import kotlin.time.toKotlinDuration
 
-internal class ValueClassUnboxConverter<T : Any>(private val valueClass: Class<T>) : StdConverter<T, Any?>() {
-    private val unboxMethod = valueClass.getDeclaredMethod("unbox-impl").apply {
-        if (!this.isAccessible) this.isAccessible = true
-    }
-
-    override fun convert(value: T): Any? = unboxMethod.invoke(value)
-
-    override fun getInputType(typeFactory: TypeFactory): JavaType = typeFactory.constructType(valueClass)
-}
-
 internal sealed class CollectionValueStrictNullChecksConverter<T : Any> : Converter<T, T> {
     protected abstract val type: JavaType
     protected abstract val paramName: String
